[fbx] Fixed handling nameless objects. Added default for missing material info. [oiRM] added start for a spec of materials [oirm_gen] Now handles file extensions regardless of casing [ostlc] String i64 and u64 now have to be explicitly called

+++ owc/src/platforms/windows/wfilemanager.cpp
@@ -51,7 +51,7 @@ bool FileManager::mkdir(String path) const {
 			DWORD error = GetLastError();
 			if (error == ERROR_ALREADY_EXISTS) continue;
 
-			return Log::error(String("Couldn't mkdir ") + current + "(" + (u64)error + ")");
+			return Log::error(String("Couldn't mkdir ") + current + "(" + (u32) error + ")");
 		}
 	}
